Upgrade Structure
Structure improvement add the `modules` path to the project to allocate modules
